Initial V2 Commit
In this commit I have scrapped the old code in order to restart with new
goals in mind. The new purpose of this program is to be used as an
adaptable software for testing algorithms on graphs in accordance for
graph theory.

I want to build it such that there are helper buttons which can do
general tasks to make it easier to use it while giving options to run
calculations at the same time.

I have started by rebuilding the basic logic of the graphs from the
original version and by adding a button system to give the user the
ability to manipulate the graph.